requirements.txt
aioredis==1.3.0
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
dnspython==2.3.0
fastapi==0.89.1
h11==0.14.0
hiredis==2.2.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2022.7.1
redis==4.5.4
redis-server==6.0.9
requests==2.28.2
sendgrid==6.9.7
sentry-sdk==1.19.1
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.22.0
tomli==2.0.1
twilio==7.16.2
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
